requirements.txt
# Telegram
python-telegram-bot>=21.0

# Async HTTP and WebSocket
httpx>=0.27.0
websockets>=12.0

# MQTT (SDCP/ElegooLink)
paho-mqtt>=2.1.0

# Network discovery (Cassini)
scapy>=2.5.0

# SDCP V3 async UDP
asyncio-dgram>=2.2.0

# HTML parsing for link metadata
beautifulsoup4>=4.12.0

# Config
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# LLM (choose one or add as needed)
# openai>=1.0.0
# ollama  # if using Ollama
